Crate ltsinfo_reduction

This is the reduction module.

It provides various functionalities for reducing labeled transition systems (LTS). The module includes different partitioning strategies, quotienting mechanisms, and other techniques to simplify and analyze LTS.

Structs

Block

A block stores a subset of the elements in a partition. It uses start, middle and end to indicate a range start..end of elements in the partition. The middle is used such that marked_split..end are the marked elements. This is useful to be able to split off new blocks cheaply.

BlockIter

BlockPartition

A partition that explicitly stores a list of blocks and their indexing into the list of elements.

BlockPartitionBuilder

CompactSignaturePair

IncomingTransitions

A struct containg information related to the incoming transitions for every state.

IndexedPartition

Defines a partition based on an explicit indexing of elements to their block number.

Signature

The type of a signature. We use sorted vectors to avoid the overhead of hash sets that might have unused values.

Traits

Partition

A trait for partition refinement algorithms that expose the block number for every state. Can be used to compute the quotient labelled transition system.

Functions

branching_bisim_signature

Returns the branching bisimulation signature for branching bisimulation sig(s, pi) = { (a, pi(t)) | s -[tau]-> s1 -> … s_n -[a]-> t in T && pi(s) = pi(s_i) && ((a != tau) || pi(s) != pi(t)) }

branching_bisim_signature_inductive

The input lts must contain no tau-cycles.

branching_bisim_signature_sorted

The input lts must contain no tau-cycles.

branching_bisim_sigref

Computes a branching bisimulation partitioning using signature refinement

branching_bisim_sigref_naive

Computes a branching bisimulation partitioning using signature refinement without dirty blocks.

combine_partition

Combines two partitions into a new partition.

has_tau_loop

Returns true iff the labelled transition system has tau-loops.

is_valid_refinement

Returns true iff the given partition is a strong bisimulation partition

preprocess_branching

Perform the preprocessing necessary for branching bisimulation with the sorted signature see branching_bisim_signature_sorted.

quotient_lts

Returns a new LTS based on the given partition.

quotient_lts_efficient

reorder_partition

Reorders the blocks of the given partition according to the given permutation.

reorder_states

scc_decomposition

Computes the strongly connected component partitioning of the given LTS.

sort_topological

Returns a topological ordering of the states of the given LTS.

strong_bisim_signature

Returns the signature for strong bisimulation sig(s, pi) = { (a, pi(t)) | s -a-> t in T }

strong_bisim_sigref

Computes a strong bisimulation partitioning using signature refinement

strong_bisim_sigref_naive

Computes a strong bisimulation partitioning using signature refinement

tau_scc_decomposition

Computes the strongly connected tau component partitioning of the given LTS.

Type Aliases

SignatureBuilder

The builder used to construct the signature.