ltsinfo_reduction

Struct BlockPartition

pub struct BlockPartition { /* private fields */ }

A partition that explicitly stores a list of blocks and their indexing into the list of elements.

Implementations

impl BlockPartition

pub fn new(num_of_elements: usize) -> BlockPartition

Create an initial partition where all the states are in a single block 0. And all the elements in the block are marked.

pub fn partition_marked_with<F>( &mut self, block_index: usize, builder: &mut BlockPartitionBuilder, partitioner: F, ) -> impl Iterator<Item = usize>
where F: FnMut(usize, &BlockPartition) -> usize,

Partition the elements of the given block into multiple new blocks based on the given partitioner; which returns a number for each marked element. Elements with the same number belong to the same block, and the returned numbers should be dense.

Returns an iterator over the new block indices, where the first element is the index of the block that was partitioned. And that block is the largest block.

pub fn split_marked( &mut self, block_index: usize, splitter: impl FnMut(usize) -> bool, )

Split the given block into two separate block based on the splitter predicate.

pub fn mark_backward_closure( &mut self, block_index: usize, incoming_transitions: &IncomingTransitions, )

Makes the marked elements closed under the silent closure of incoming tau-transitions within the current block. Also guarantees that the elements are topologically sorted.

pub fn swap_blocks(&mut self, left_index: usize, right_index: usize)

Swaps the given blocks given by the indices.

pub fn mark_element(&mut self, element: usize)

Marks the given element, such that it is returned by iter_marked.

pub fn is_element_marked(&self, element: usize) -> bool

Returns true iff the given element has already been marked.

pub fn block(&self, block_index: usize) -> &Block

Return a reference to the given block.

pub fn num_of_blocks(&self) -> usize

Returns the number of blocks in the partition.

pub fn iter_block(&self, block_index: usize) -> BlockIter<'_>

Returns an iterator over the elements of a given block.