ltsinfo_reduction

Trait Partition

pub trait Partition {
    // Required methods
    fn block_number(&self, state_index: usize) -> usize;
    fn num_of_blocks(&self) -> usize;
    fn len(&self) -> usize;

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn equal(&self, other: &impl Partition) -> bool { ... }
}

A trait for partition refinement algorithms that expose the block number for every state. Can be used to compute the quotient labelled transition system.

The invariants are that the union of all blocks is the original set, and that each block contains distinct elements

Required Methods

fn block_number(&self, state_index: usize) -> usize

Returns the block number for the given state.

fn num_of_blocks(&self) -> usize

Returns the number of blocks in the partition.

fn len(&self) -> usize

Returns the number of elements in the partition.

Provided Methods

fn is_empty(&self) -> bool

Returns whether the partition is empty.

fn equal(&self, other: &impl Partition) -> bool

Returns true iff the partitions are equal, runs in O(n^2)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.